보스 구조 및 패턴 기초

+++ Assets/Scripts/Data/Stat.cs
@@ -5,9 +5,9 @@ using UnityEngine;
 public class Stat : MonoBehaviour
 {
     [SerializeField]
-    protected int _hp;
+    protected float _hp;
     [SerializeField]
-    protected int _maxHp;
+    protected float _maxHp;
     [SerializeField]
     protected int _attack;
     [SerializeField]
@@ -20,8 +20,8 @@ public class Stat : MonoBehaviour
     [SerializeField]
     protected float _attackSpeed;
 
-    public int Hp { get { return _hp; } set { _hp = value; } }
-    public int MaxHp { get { return _maxHp; } set { _maxHp = value; } }
+    public float Hp { get { return _hp; } set { _hp = value; } }
+    public float MaxHp { get { return _maxHp; } set { _maxHp = value; } }
     public int Attack { get { return _attack; } set { _attack = value; } }
     public int Defense { get { return _defense; } set { _defense = value; } }
     public float MoveSpeed { get { return _moveSpeed; } set { _moveSpeed = value; } }
